+++ source/javascripts/pages/TriggersPage/components/LegacyTriggers/AddPrTriggerDialog.tsx
@@ -18,7 +18,7 @@ import { Controller, FormProvider, useFieldArray, useForm } from 'react-hook-for
 
 import { FormItems, LegacyPrConditionType, TriggerItem } from '../TriggersPage/TriggersPage.types';
 import { checkIsConditionsUsed } from '../TriggersPage/TriggersPage.utils';
-import ConditionCard from '../SelectiveTriggers/ConditionCard';
+import ConditionCard from '../TargetBasedTriggers/ConditionCard';
 
 type DialogProps = {
   currentTriggers: TriggerItem[];
