+++ source/javascripts/pages/TriggersPage/components/TargetBasedTriggers/ConditionCard.tsx
