+++ source/javascripts/pages/TriggersPage/components/TargetBasedTriggers/RegexCheckbox.tsx
